+++ test/glue/pulp/repository_test.rb
@@ -11,13 +11,13 @@
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
 
 require 'minitest_helper'
-
+require './test/support/pulp/task_support'
 
 class GluePulpRepoTestBase < MiniTest::Rails::ActiveSupport::TestCase
   extend ActiveRecord::TestFixtures
+  include TaskSupport
 
   fixtures :all
-  @@admin = nil
 
   def self.before_suite
     @loaded_fixtures = load_fixtures
@@ -29,34 +29,11 @@ class GluePulpRepoTestBase < MiniTest::Rails::ActiveSupport::TestCase
     disable_glue_layers(services, models, true)
 
     @@admin = User.find(@loaded_fixtures['users']['admin']['id'])
-  end
+    User.current = @@admin
 
-  def self.wait_on_tasks(task_list)
-    VCR.use_cassette('glue_pulp_repo_tasks', :erb => true, :match_requests_on => [:path, :method]) do
-      task_list.each do |task|
-        while !(['finished', 'error', 'timed_out', 'canceled', 'reset', 'success'].include?(task['state'])) do
-          task = PulpSyncStatus.pulp_task(Katello.pulp_server.resources.task.poll(task["task_id"]))
-          sleep 0.5 # do not overload backend engines
-        end
-      end
-    end
-  end
-
-  def self.wait_on_task(task)
-    VCR.use_cassette('glue_pulp_repo_tasks', :erb => true, :match_requests_on => [:path, :method]) do
-      while !(['finished', 'error', 'timed_out', 'canceled', 'reset', 'success'].include?(task['state'])) do
-        task = PulpSyncStatus.pulp_task(Katello.pulp_server.resources.task.poll(task["uuid"]))
-        sleep 0.5 # do not overload backend engines
-      end
-    end
-  end
-
-  def setup
-    VCR.insert_cassette('glue_pulp_repo', :match_requests_on => [:path, :params, :method])
-  end
-
-  def teardown
-    VCR.eject_cassette
+    @@fedora_17_x86_64 = Repository.find(@loaded_fixtures['repositories']['fedora_17_x86_64']['id'])
+    @@fedora_17_x86_64.relative_path = '/test_path/'
+    @@fedora_17_x86_64.feed = "file:///var/www/test_repos/zoo"
   end
 
 end
@@ -66,9 +43,12 @@ class GluePulpRepoTestCreateDestroy < GluePulpRepoTestBase
 
   def setup
     super
-    @fedora_17_x86_64 = Repository.find(repositories(:fedora_17_x86_64).id)
-    @fedora_17_x86_64.relative_path = '/test_path/'
-    @fedora_17_x86_64.feed = "file://#{File.expand_path(File.dirname(__FILE__))}".gsub("glue/pulp", "fixtures/zoo5")
+    @fedora_17_x86_64 = @@fedora_17_x86_64
+    VCR.insert_cassette('pulp/repository/create')
+  end
+
+  def teardown
+    VCR.eject_cassette
   end
 
   def test_create_pulp_repo
@@ -83,54 +63,27 @@ class GluePulpRepoTestCreateDestroy < GluePulpRepoTestBase
 
 end
 
-
 class GluePulpRepoTest < GluePulpRepoTestBase
-
-  @@fedora_17_x86_64 = nil
 
   def self.before_suite
     super
-    @@fedora_17_x86_64 = Repository.find(@loaded_fixtures['repositories']['fedora_17_x86_64']['id'])
-    @@fedora_17_x86_64.relative_path = '/test_path/'
-    @@fedora_17_x86_64.feed = "file://#{File.expand_path(File.dirname(__FILE__))}".gsub("glue/pulp", "fixtures/zoo5")
-
-    VCR.use_cassette('glue_pulp_repo_helper') do
-      @@fedora_17_x86_64.create_pulp_repo
-    end
+    VCR.insert_cassette('pulp/repository/repository')
+    @@fedora_17_x86_64.create_pulp_repo
   end
 
   def self.after_suite
-    VCR.use_cassette('glue_pulp_repo_helper') do
-      @@fedora_17_x86_64.destroy_repo
-    end
+    @@fedora_17_x86_64.destroy_repo
+    VCR.eject_cassette
   end
 
   def setup
     super
-    User.current = @@admin
     @fedora_17_x86_64 = @@fedora_17_x86_64
     @fedora_17_x86_64.relative_path = '/test_path/'
   end
 
-  def test_sync
-    task_list = @fedora_17_x86_64.sync
-
-    refute_empty    task_list
-    assert_kind_of  PulpSyncStatus, task_list.first
-
-    @task = task_list.first
-    self.class.wait_on_task(@task)
-  end
-
-  def test_set_sync_schedule
-    VCR.use_cassette('glue_pulp_repo_sync_schedule', :match_requests_on => [:path, :params, :method]) do
-      assert @fedora_17_x86_64.set_sync_schedule(Time.now.advance(:years => 1).iso8601 << "/P1D")
-    end
-  end
-
-  def test_cancel_sync
-    #@@fedora_17_x86_64.sync
-    #assert @@fedora_17_x86_64.cancel_sync
+  def test_delete_orphaned_content
+    assert Repository.delete_orphaned_content.is_a?(Hash)
   end
 
   def test_relative_path
@@ -143,44 +96,143 @@ class GluePulpRepoTest < GluePulpRepoTestBase
     refute_equal '/test_path/', @fedora_17_x86_64.relative_path
   end
 
-  def test_generate_distributor
-    assert_kind_of Runcible::Models::YumDistributor, @fedora_17_x86_64.generate_distributor
+  def test_generate_distributors
+    dists = @fedora_17_x86_64.generate_distributors
+    refute_empty dists.select{|d| d.is_a? Runcible::Models::YumDistributor}
+    refute_empty dists.select{|d| d.is_a? Runcible::Models::YumCloneDistributor}
   end
 
   def test_populate_from
     assert @fedora_17_x86_64.populate_from({ @fedora_17_x86_64.pulp_id => {} })
   end
 
+  def test_sync
+    task_list = @fedora_17_x86_64.sync
+    refute_empty    task_list
+    assert_kind_of  PulpSyncStatus, task_list.first
+
+    TaskSupport.wait_on_tasks(task_list)
+  end
+
+  def test_set_sync_schedule
+    time = "2013-08-01T00:00:00-04:00/P1D"
+    assert @fedora_17_x86_64.set_sync_schedule(time)
+  end
+
 end
 
 
-class GluePulpRepoRequiresSyncTest < GluePulpRepoTestBase
-
-  i_suck_and_my_tests_are_order_dependent!
-
-  @@fedora_17_x86_64 = nil
-  @@fedora_17_x86_64_dev = nil
+class GluePulpRepoAfterSyncTest < GluePulpRepoTestBase
 
   def self.before_suite
     super
-    User.current = @@admin
-    @@fedora_17_x86_64 = Repository.find(@loaded_fixtures['repositories']['fedora_17_x86_64']['id'])
-    @@fedora_17_x86_64_dev = Repository.find(@loaded_fixtures['repositories']['fedora_17_x86_64_dev']['id'])
-
-    @@fedora_17_x86_64.relative_path = '/test_path/'
-    @@fedora_17_x86_64.feed = "file://#{File.expand_path(File.dirname(__FILE__))}".gsub("glue/pulp", "fixtures/zoo5")
-
-    VCR.use_cassette('glue_pulp_repo_helper') do
-      @@fedora_17_x86_64.create_pulp_repo
-      task = @@fedora_17_x86_64.sync.first
-      wait_on_task(task)
-    end
+    VCR.insert_cassette('pulp/repository/after_sync')
+    @@fedora_17_x86_64.create_pulp_repo
   end
 
   def self.after_suite
-    VCR.use_cassette('glue_pulp_repo_helper') do
-      @@fedora_17_x86_64.destroy_repo
+    @@fedora_17_x86_64.destroy_repo
+    VCR.eject_cassette
+  end
+
+  def setup
+    super
+    @fedora_17_x86_64 = @@fedora_17_x86_64
+    @fedora_17_x86_64.relative_path = '/test_path/'
+  end
+
+  def test_handle_sync_complete_task
+    mock_notifier = Minitest::Mock.new
+    mock_notifier.expect(:success, nil)
+
+    task_list = @fedora_17_x86_64.sync
+    TaskSupport.wait_on_tasks(task_list)
+
+    @fedora_17_x86_64.handle_sync_complete_task(task_list.first.uuid, mock_notifier)
+
+    assert PulpTaskStatus.where(:uuid => task_list.first.uuid).length > 0
+  end
+
+end
+
+
+class GluePulpChangeFeedTest < GluePulpRepoTestBase
+
+  def setup
+    VCR.insert_cassette('pulp/repository/feed_change')
+    @@fedora_17_x86_64.create_pulp_repo
+  end
+
+  def test_feed_change
+    new_feed = "http://foo.com/foo"
+    @@fedora_17_x86_64.feed = new_feed
+    @@fedora_17_x86_64.save!
+    pulps_feed = Repository.find(@@fedora_17_x86_64.id).pulp_repo_facts['importers'].first['config']['feed']
+    assert_equal new_feed, pulps_feed
+  end
+
+  def teardown
+    @@fedora_17_x86_64.destroy_repo
+    VCR.eject_cassette
+  end
+end
+
+class GluePulpPuppetRepoTest < GluePulpRepoTestBase
+
+  @@p_forge = nil
+
+  def self.before_suite
+    super
+    VCR.insert_cassette('pulp/repository/puppet')
+    @@p_forge = Repository.find(@loaded_fixtures['repositories']['p_forge']['id'])
+    @@p_forge.relative_path = '/test_path/'
+    @@p_forge.feed = "http://davidd.fedorapeople.org/repos/random_puppet/"
+    @@p_forge.create_pulp_repo
+  end
+
+  def self.after_suite
+    @@p_forge.destroy_repo
+    VCR.eject_cassette
+  end
+
+  def setup
+    super
+    User.current = @@admin
+    @p_forge = @@p_forge
+    @p_forge.relative_path = '/test_path/'
+  end
+
+  def test_generate_distributors
+    refute_nil @@p_forge.find_distributor
+  end
+
+  def test_upload_puppet_module
+    Repository.any_instance.expects(:trigger_contents_changed).with() do |options|
+      options[:wait] == false && options[:reindex] == false && options[:index_units].size > 0
     end
+
+    @filepath = File.join(Rails.root, "test/fixtures/puppet/puppetlabs-ntp-2.0.1.tar.gz")
+    @p_forge.upload_content(@filepath)
+    assert_includes @p_forge.puppet_modules.map(&:name), "ntp"
+  end
+end
+
+
+
+class GluePulpRepoContentsTest < GluePulpRepoTestBase
+
+  def self.before_suite
+    super
+    VCR.insert_cassette('pulp/repository/contents')
+
+    @@fedora_17_x86_64.create_pulp_repo
+    task_list = @@fedora_17_x86_64.sync
+    TaskSupport.wait_on_tasks(task_list)
+  end
+
+  def self.after_suite
+    @@fedora_17_x86_64.destroy_repo
+    VCR.eject_cassette
   end
 
   def test_last_sync
@@ -208,55 +260,39 @@ class GluePulpRepoRequiresSyncTest < GluePulpRepoTestBase
   end
 
   def test_sync_finish
-    VCR.use_cassette('glue_pulp_repo_sync_status', :match_requests_on => [:path, :params, :method]) do
-      refute_nil @@fedora_17_x86_64.sync_finish
-    end
+    refute_nil @@fedora_17_x86_64.sync_finish
   end
 
   def test_sync_start
-    VCR.use_cassette('glue_pulp_repo_sync_status', :match_requests_on => [:path, :params, :method]) do
-      refute_nil @@fedora_17_x86_64.sync_start
-    end
+    refute_nil @@fedora_17_x86_64.sync_start
   end
 
   def test_packages
-    VCR.use_cassette('glue_pulp_repo_units', :match_requests_on => [:body_json, :path, :method]) do
-      refute_empty @@fedora_17_x86_64.packages.select { |package| package.name == 'elephant' }
-    end
+    refute_empty @@fedora_17_x86_64.packages.select { |package| package.name == 'elephant' }
   end
 
   def test_has_package?
-    VCR.use_cassette('glue_pulp_repo_units', :match_requests_on => [:body_json, :path, :method]) do
-      pkg_id = @@fedora_17_x86_64.packages.sort_by(&:id).first.id
-      assert @@fedora_17_x86_64.has_package?(pkg_id)
-    end
+    pkg_id = @@fedora_17_x86_64.packages.sort_by(&:id).first.id
+    assert @@fedora_17_x86_64.has_package?(pkg_id)
   end
 
   def test_errata
-    VCR.use_cassette('glue_pulp_repo_units', :match_requests_on => [:body_json, :path, :method]) do
-      refute_empty @@fedora_17_x86_64.errata.select { |errata| errata.errata_id == "RHEA-2010:0002" }
-    end
+    refute_empty @@fedora_17_x86_64.errata.select { |errata| errata.errata_id == "RHEA-2010:0002" }
   end
 
   def test_has_erratum?
-    VCR.use_cassette('glue_pulp_repo_units', :match_requests_on => [:body_json, :path, :method]) do
-      e_id = @@fedora_17_x86_64.errata.first.errata_id
-      assert @@fedora_17_x86_64.has_erratum?(e_id)
-    end
+    e_id = @@fedora_17_x86_64.errata.first.errata_id
+    assert @@fedora_17_x86_64.has_erratum?(e_id)
   end
 
   def test_distributions
-    VCR.use_cassette('glue_pulp_repo_units', :match_requests_on => [:body_json, :path, :method]) do
-      distributions = @@fedora_17_x86_64.distributions
+    distributions = @@fedora_17_x86_64.distributions
 
-      refute_empty distributions.select { |distribution| distribution.id == "ks-Test Family-TestVariant-16-x86_64" }
-    end
+    refute_empty distributions.select { |distribution| distribution.id == "ks-Test Family-TestVariant-16-x86_64" }
   end
 
   def test_has_distribution?
-    VCR.use_cassette('glue_pulp_repo_units', :match_requests_on => [:body_json, :path, :method]) do
-      assert @@fedora_17_x86_64.has_distribution?("ks-Test Family-TestVariant-16-x86_64")
-    end
+    assert @@fedora_17_x86_64.has_distribution?("ks-Test Family-TestVariant-16-x86_64")
   end
 
   def test_find_packages_by_name
@@ -272,19 +308,46 @@ class GluePulpRepoRequiresSyncTest < GluePulpRepoTestBase
   end
 
   def test_package_groups
-    VCR.use_cassette('glue_pulp_repo_units', :match_requests_on => [:body_json, :path, :method]) do
-      package_groups = @@fedora_17_x86_64.package_groups
+    package_groups = @@fedora_17_x86_64.package_groups
 
-      refute_empty package_groups.select { |group| group.name == 'mammal' }
-    end
+    refute_empty package_groups.select { |group| group.name == 'mammal' }
   end
 
   def test_package_group_categories
-    VCR.use_cassette('glue_pulp_repo_units', :match_requests_on => [:body_json, :path, :method]) do
-      categories = @@fedora_17_x86_64.package_group_categories
+    categories = @@fedora_17_x86_64.package_group_categories
 
-      refute_empty categories.select { |category| category['name'] == 'all' }
-    end
+    refute_empty categories.select { |category| category['name'] == 'all' }
+  end
+
+  def test_trigger_contents_changed_index_units
+    Katello.config.stubs(:use_elasticsearch).returns(:true)
+    pkg = @@fedora_17_x86_64.find_packages_by_nvre('elephant', '0.3', '0.8', '0')[0]
+    @@fedora_17_x86_64.expects(:generate_metadata).returns([])
+    Package.expects(:index_packages).with([pkg['_id']])
+
+    unit = {:checksumtype => pkg['checksumtype'], :checksum => pkg['checksum'] }
+    @@fedora_17_x86_64.trigger_contents_changed(:publish => true, :reindex => false, :index_units => [unit])
+  end
+
+end
+
+
+class GluePulpRepoOperationsTest < GluePulpRepoTestBase
+
+  def self.before_suite
+    super
+    VCR.insert_cassette('pulp/repository/operations')
+
+    @@fedora_17_x86_64_dev = Repository.find(@loaded_fixtures['repositories']['fedora_17_x86_64_dev']['id'])
+
+    @@fedora_17_x86_64.create_pulp_repo
+    task_list = @@fedora_17_x86_64.sync
+    TaskSupport.wait_on_tasks(task_list)
+  end
+
+  def self.after_suite
+    @@fedora_17_x86_64.destroy_repo
+    VCR.eject_cassette
   end
 
   def test_create_clone
@@ -305,7 +368,7 @@ class GluePulpRepoRequiresSyncTest < GluePulpRepoTestBase
     task_list = @@fedora_17_x86_64.clone_contents(@@fedora_17_x86_64_dev)
     assert_equal 5, task_list.length
 
-    self.class.wait_on_tasks(task_list)
+    TaskSupport.wait_on_tasks(task_list)
   ensure
     @@fedora_17_x86_64_dev.destroy_repo
   end
@@ -316,7 +379,7 @@ class GluePulpRepoRequiresSyncTest < GluePulpRepoTestBase
 
     task_list = @@fedora_17_x86_64.promote(library, staging)
     assert_equal 5, task_list.length
-    self.class.wait_on_tasks(task_list)
+    TaskSupport.wait_on_tasks(task_list)
 
     clone_id = @@fedora_17_x86_64.clone_id(staging, staging.default_content_view)
     cloned_repo = Repository.where(:pulp_id => clone_id).first
@@ -328,44 +391,35 @@ end
 
 class GluePulpRepoRequiresEmptyPromoteTest < GluePulpRepoTestBase
 
-  @@fedora_17_x86_64  = nil
-  @@cloned_repo       = nil
-  @@staging           = nil
-  @@library           = nil
+
+  @@cloned_repo = nil
 
   def self.before_suite
     super
-    User.current = @@admin
-
-    @@fedora_17_x86_64 = Repository.find(@loaded_fixtures['repositories']['fedora_17_x86_64']['id'])
-    @@fedora_17_x86_64.relative_path = '/test_path/'
-    @@fedora_17_x86_64.feed = "file://#{File.expand_path(File.dirname(__FILE__))}".gsub("glue/pulp", "fixtures/zoo5")
+    VCR.insert_cassette('pulp/repository/add_contents')
 
     @@library = KTEnvironment.find(@loaded_fixtures['environments']['library']['id'])
     @@staging = KTEnvironment.find(@loaded_fixtures['environments']['staging']['id'])
 
-    VCR.use_cassette('glue_pulp_repo_helper') do
-      #clone_id = @@fedora_17_x86_64.clone_id(@@staging)
-      #Runcible::Resources::Repository.delete(clone_id)
-      ##@@fedora_17_x86_64.destroy_repo
-      @@fedora_17_x86_64.create_pulp_repo
-      task_list = @@fedora_17_x86_64.promote(@@library, @@staging)
-      wait_on_tasks(task_list)
+    @@fedora_17_x86_64.create_pulp_repo
+    task_list = @@fedora_17_x86_64.sync
+    TaskSupport.wait_on_tasks(task_list)
 
-      task = @@fedora_17_x86_64.sync.first
-      wait_on_task(task)
+    #clone_id = @@fedora_17_x86_64.clone_id(@@staging, @@staging.default_content_view)
+    #Runcible::Extensions::Repository.delete(clone_id)
 
 
-      clone_id = @@fedora_17_x86_64.clone_id(@@staging, @@staging.default_content_view)
-      @@cloned_repo = Repository.where(:pulp_id => clone_id).first
-    end
+    task = @@fedora_17_x86_64.promote(@@library, @@staging)
+    TaskSupport.wait_on_tasks(task)
+
+    clone_id = @@fedora_17_x86_64.clone_id(@@staging, @@staging.default_content_view)
+    @@cloned_repo = Repository.where(:pulp_id => clone_id).first
   end
 
   def self.after_suite
-    VCR.use_cassette('glue_pulp_repo_helper') do
-      @@cloned_repo.destroy if @@cloned_repo
-      @@fedora_17_x86_64.destroy_repo if @@fedora_17_x86_64
-    end
+    @@cloned_repo.destroy if @@cloned_repo
+    @@fedora_17_x86_64.destroy_repo if @@fedora_17_x86_64
+    VCR.eject_cassette
   end
 
   def test_add_packages
@@ -394,33 +448,26 @@ class GluePulpRepoRequiresSyncAndPromoteTest < GluePulpRepoTestBase
 
   def self.before_suite
     super
-    User.current = @@admin
-
-    @@fedora_17_x86_64 = Repository.find(@loaded_fixtures['repositories']['fedora_17_x86_64']['id'])
-    @@fedora_17_x86_64.relative_path = '/test_path/'
-    @@fedora_17_x86_64.feed = "file://#{File.expand_path(File.dirname(__FILE__))}".gsub("glue/pulp", "fixtures/zoo5")
+    VCR.insert_cassette('pulp/repository/remove_contents')
 
     @@library = KTEnvironment.find(@loaded_fixtures['environments']['library']['id'])
     @@staging = KTEnvironment.find(@loaded_fixtures['environments']['staging']['id'])
 
-    VCR.use_cassette('glue_pulp_repo_helper') do
-      @@fedora_17_x86_64.create_pulp_repo
-      task = @@fedora_17_x86_64.sync.first
-      wait_on_task(task)
+    @@fedora_17_x86_64.create_pulp_repo
+    task_list = @@fedora_17_x86_64.sync
+    TaskSupport.wait_on_tasks(task_list)
 
-      task_list = @@fedora_17_x86_64.promote(@@library, @@staging)
-      wait_on_tasks(task_list)
+    task_list = @@fedora_17_x86_64.promote(@@library, @@staging)
+    TaskSupport.wait_on_tasks(task_list)
 
-      clone_id = @@fedora_17_x86_64.clone_id(@@staging, @@staging.default_content_view)
-      @@cloned_repo = Repository.where(:pulp_id => clone_id).first
-    end
+    clone_id = @@fedora_17_x86_64.clone_id(@@staging, @@staging.default_content_view)
+    @@cloned_repo = Repository.where(:pulp_id => clone_id).first
   end
 
   def self.after_suite
-    VCR.use_cassette('glue_pulp_repo_helper') do
-      @@fedora_17_x86_64.destroy_repo
-      @@cloned_repo.destroy
-    end
+    @@cloned_repo.destroy if @@cloned_repo
+    @@fedora_17_x86_64.destroy_repo if @@fedora_17_x86_64
+    VCR.eject_cassette
   end
 
   def test_delete_packages
